(maint) Refactoring
