+++ src/Tests/Lanceur.Tests/BusinessLogic/AliasManagementShould.cs
@@ -4,11 +4,10 @@ using FluentAssertions;
 using FluentAssertions.Execution;
 using Lanceur.Core.Models;
 using Lanceur.Core.Services;
-using Lanceur.Infra.SQLite;
 using Lanceur.Infra.SQLite.DataAccess;
 using Lanceur.Infra.SQLite.DbActions;
 using Lanceur.Infra.SQLite.Repositories;
-using Lanceur.Tests.SQLite;
+using Lanceur.Tests.Tooling;
 using Lanceur.Ui.Core.Utils;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
