+++ src/Tests/Lanceur.Tests/BusinessLogic/MacroShould.cs
@@ -6,9 +6,7 @@ using Lanceur.Core.Models;
 using Lanceur.Core.Repositories;
 using Lanceur.Core.Services;
 using Lanceur.Infra.Managers;
-using Lanceur.Infra.SQLite;
 using Lanceur.Infra.Utils;
-using Lanceur.Tests.SQLite;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using System.Reflection;
@@ -18,11 +16,11 @@ using Lanceur.Infra.Macros;
 using Lanceur.Infra.SQLite.DataAccess;
 using Lanceur.Infra.SQLite.DbActions;
 using Lanceur.Infra.SQLite.Repositories;
+using Lanceur.Tests.Tooling;
 using Lanceur.Tests.Tooling.Extensions;
 using Lanceur.Tests.Tooling.Macros;
 using Lanceur.Ui.Core.Utils;
 using Microsoft.Extensions.DependencyInjection;
-using Serilog.Core;
 using Xunit;
 using Xunit.Abstractions;
