+++ src/Tests/Lanceur.Tests/Tooling/ConnectionStringFactory.cs
