+++ src/Tests/Lanceur.Tests/Tooling/DebugStoreLoader.cs
