+++ src/Tests/Lanceur.Tests/Tooling/Extensions/ServiceProviderExtensions.cs
