+++ src/Tests/Lanceur.Tests/Tooling/Extensions/TestOutputHelperExtensions.cs
