+++ src/Tests/Lanceur.Tests/Tooling/Logging/BaseTestOutputHelperDecorator.cs
