+++ src/Tests/Lanceur.Tests/Tooling/Logging/MicrosoftLoggingLoggerFactory.cs
