+++ src/Tests/Lanceur.Tests/Tooling/Logging/TestOutputHelperMixin.cs
