+++ src/Tests/Lanceur.Tests/Tooling/Macros/MultiMacroTest.cs
