+++ src/Tests/Lanceur.Tests/Tooling/ObjectExtensions.cs
