+++ src/Tests/Lanceur.Tests/Tooling/ReservedAliases/ExecutableTestAlias.cs
