+++ src/Tests/Lanceur.Tests/Tooling/ReservedAliases/NotExecutableTestAlias.cs
