+++ src/Tests/Lanceur.Tests/Tooling/SQL/SqlBuilder.cs
