+++ src/Tests/Lanceur.Tests/Tooling/TestOutputHelperMixin.cs
