+++ src/Tests/Lanceur.Tests/UseCases/AliasUseCases.cs
@@ -12,8 +12,7 @@ using Lanceur.Infra.Managers;
 using Lanceur.Infra.Services;
 using Lanceur.Infra.SQLite.DbActions;
 using Lanceur.Infra.Stores;
-using Lanceur.Tests._Tooling.StateTesters;
-using Lanceur.Tests.SQLite;
+using Lanceur.Tests.Tooling.StateTesters;
 using Lanceur.Tests.Tooling;
 using Lanceur.Tests.Tooling.Extensions;
 using Lanceur.Tests.Tooling.SQL;
