+++ src/Tests/Lanceur.Tests/ViewModels/MainViewModelShould.cs
@@ -10,10 +10,9 @@ using Lanceur.Core.Services;
 using Lanceur.Core.Stores;
 using Lanceur.Infra.Managers;
 using Lanceur.Infra.Services;
-using Lanceur.Infra.SQLite;
 using Lanceur.Infra.SQLite.DbActions;
 using Lanceur.Infra.Stores;
-using Lanceur.Tests.SQLite;
+using Lanceur.Tests.Tooling;
 using Lanceur.Tests.Tooling.Extensions;
 using Lanceur.Tests.Tooling.SQL;
 using Lanceur.Ui.Core.Utils;
